requirements.txt
PyQt5
PyYAML
llama-cpp-python